third_party/blink/web_tests/external/wpt/websockets/interfaces/WebSocket/close/close-nested_default-expected.txt

This is a testharness.js-based test.
FAIL WebSockets: close() in close event handler assert_equals: expected 2 but got 3
Harness: the test ran to completion.
